test(cli): split list inference suite
